refactor edit, tags and libraries
