+++ src/edit_asset/__init__.py
@@ -1,0 +1,1 @@
+"""Edit asset module"""
