+++ src/edit_asset/operators.py
@@ -1,14 +1,14 @@
-"""Edit assets module."""
+"""Edit assets operators."""
 import uuid
 
 import bpy
 
-from .async_functions import delete_asset, edit_asset, edit_view
-from .export_data import get_edit_data
-from .upload import get_edit_props
+from ...config import HANA3D_DESCRIPTION, HANA3D_NAME
 from ..async_loop.async_mixin import AsyncModalOperatorMixin
 from ..ui.main import UI
-from ...config import HANA3D_DESCRIPTION, HANA3D_NAME
+from .async_functions import delete_asset, edit_asset, edit_view
+from .edit import get_edit_props
+from .export_data import get_edit_data
 
 HANA3D_EXPORT_DATA_FILE = f'{HANA3D_NAME}_data.json'
