+++ src/ui/operators/asset_bar.py
@@ -8,11 +8,6 @@ from bpy.props import BoolProperty, StringProperty
 from bpy_extras import view3d_utils
 from mathutils import Vector
 
-from ..callbacks.asset_bar import draw_callback2d, draw_callback3d
-from ..main import UI
-from ...preferences.preferences import Preferences
-from ...search import search
-from ...upload import upload
 from .... import utils
 from ....config import (
     HANA3D_DESCRIPTION,
@@ -21,6 +16,12 @@ from ....config import (
     HANA3D_UI,
 )
 from ....report_tools import execute_wrapper
+from ...edit_asset.edit import set_edit_props
+from ...preferences.preferences import Preferences
+from ...search import search
+from ...upload import upload
+from ..callbacks.asset_bar import draw_callback2d, draw_callback3d
+from ..main import UI
 
 
 def get_asset_under_mouse(mousex: float, mousey: float) -> int:
@@ -308,8 +309,9 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
             if active_obj is None:
                 return
 
-            model = ui_props.asset_type == 'MODEL'
-            material = ui_props.asset_type == 'MATERIAL' and active_obj.active_material is not None
+            model = ui_props.asset_type_search == 'MODEL'
+            if active_obj.active_material is not None:
+                material = ui_props.asset_type_search == 'MATERIAL'
             if model or material:
                 props = upload.get_upload_props()
                 asset_data = {
@@ -327,7 +329,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
         has_hit, *_ = raycast
 
         # MODELS can be dragged on scene floor
-        if not has_hit and ui_props.asset_type == 'MODEL':
+        if not has_hit and ui_props.asset_type_search == 'MODEL':
             raycast = floor_raycast(context, mx, my)
 
         ui_props.has_hit = raycast[0]
@@ -439,7 +441,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
 
             if ui_props.drag_init:
                 ui_props.drag_length += 1
-                if ui_props.drag_length > 0:
+                if ui_props.drag_length > 5:
                     ui_props.dragging = True
                     ui_props.drag_init = False
 
@@ -488,7 +490,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
             elif ui_props.dragging and mouse_in_region(region, mx, my):
                 self._raycast_update_props(ui_props, context, mx, my)
 
-            if ui_props.has_hit and ui_props.asset_type == 'MODEL':
+            if ui_props.has_hit and ui_props.asset_type_search == 'MODEL':
                 # this condition is here to fix a bug for a scene
                 # submitted by a user, so this situation shouldn't
                 # happen anymore, but there might exists scenes
@@ -511,14 +513,13 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
             my = event.mouse_y - region.y
 
         if event.type == 'LEFTMOUSE':
-
             region = self.region
             mx = event.mouse_x - region.x
             my = event.mouse_y - region.y
 
             ui_props = getattr(context.window_manager, HANA3D_UI)
             if event.value == 'PRESS' and ui_props.active_index > -1:
-                if ui_props.asset_type in {'MODEL', 'MATERIAL'}:
+                if ui_props.asset_type_search in {'MODEL', 'MATERIAL'}:
                     # check if asset is locked and let the user know in that case
                     asset_search_index = ui_props.active_index
                     asset_data = search_results[asset_search_index]
@@ -527,7 +528,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                     bpy.context.window.cursor_set('NONE')
                     ui_props.draw_tooltip = False
                     ui_props.drag_length = 0
-                elif ui_props.asset_type == 'SCENE':
+                elif ui_props.asset_type_search == 'SCENE':
                     ui_props.drag_init = True
                     bpy.context.window.cursor_set('NONE')
                     ui_props.draw_tooltip = False
@@ -564,7 +565,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                     # raycast here
                     ui_props.active_index = -3
 
-                    if ui_props.asset_type == 'MODEL':
+                    if ui_props.asset_type_search == 'MODEL':
                         raycast = self._raycast_update_props(ui_props, context, mx, my)
 
                         if not ui_props.has_hit:
@@ -576,7 +577,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                             target_object = obj_hit.name
                         target_slot = ''
 
-                    elif ui_props.asset_type == 'MATERIAL':
+                    elif ui_props.asset_type_search == 'MATERIAL':
                         raycast = mouse_raycast(context, mx, my)
 
                         ui_props.has_hit = raycast[0]
@@ -622,21 +623,17 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                 # Click interaction
                 else:
                     asset_search_index = get_asset_under_mouse(mx, my)
+                    bpy.context.view_layer.object.active = None
+                    set_edit_props(asset_search_index)
 
-                    if ui_props.asset_type in {'MATERIAL', 'MODEL'}:  # noqa: WPS220
-                        ao = bpy.context.active_object
-                        if ao is not None and not ao.is_library_indirect:
-                            target_object = bpy.context.active_object.name
-                            target_slot = bpy.context.active_object.active_material_index
-                        else:
-                            target_object = ''
-                            target_slot = ''
+                    return {'RUNNING_MODAL'}
+
                 # FIRST START SEARCH
 
                 if asset_search_index == -3:
                     return {'RUNNING_MODAL'}
                 if asset_search_index > -3:
-                    if ui_props.asset_type == 'MATERIAL':
+                    if ui_props.asset_type_search == 'MATERIAL':
                         if target_object != '':
                             # position is for downloader:
                             loc = ui_props.snapped_location
@@ -646,7 +643,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                             download_op = getattr(bpy.ops.scene, f'{HANA3D_NAME}_download')
                             download_op(
                                 True,  # noqa: WPS425
-                                asset_type=ui_props.asset_type,
+                                asset_type=ui_props.asset_type_search,
                                 asset_index=asset_search_index,
                                 model_location=loc,
                                 model_rotation=rotation,
@@ -654,7 +651,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                                 material_target_slot=target_slot,
                             )
 
-                    elif ui_props.asset_type == 'MODEL':
+                    elif ui_props.asset_type_search == 'MODEL':
                         if ui_props.has_hit and ui_props.dragging:
                             loc = ui_props.snapped_location
                             rotation = ui_props.snapped_rotation
@@ -665,7 +662,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                         download_op = getattr(bpy.ops.scene, f'{HANA3D_NAME}_download')
                         download_op(
                             True,  # noqa: WPS425
-                            asset_type=ui_props.asset_type,
+                            asset_type=ui_props.asset_type_search,
                             asset_index=asset_search_index,
                             model_location=loc,
                             model_rotation=rotation,
@@ -675,7 +672,7 @@ class AssetBarOperator(bpy.types.Operator):  # noqa: WPS338, WPS214
                     else:
                         download_op = getattr(bpy.ops.scene, f'{HANA3D_NAME}_download')
                         download_op(
-                            asset_type=ui_props.asset_type,
+                            asset_type=ui_props.asset_type_search,
                             asset_index=asset_search_index,
                         )
